add some register operation successful solve the .inst bug in reset.c
bug description: In assmber '.inst' must following with const expression, but before we use '.inst', we don't know the value of expression. So we use  macros to place variable value with const expression.
But there is some edition must to do. Main work is to redifined MRS_S
and MSR_S.